requirements.txt
apsw==3.46.0.1
aptdaemon==2.0.2
argcomplete==3.5.3
asttokens==3.0.0
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
bcc==0.30.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brlapi==0.8.6
Brotli==1.1.0
brotlicffi==1.1.0.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
click==8.1.8
cloud-init==25.1.4
command-not-found==0.3
configobj==5.0.9
contourpy==1.3.1
cryptography==43.0.0
css-parser==1.0.10
cssselect==1.3.0
cssselect2==0.8.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
distro-info==1.13
dnspython==2.7.0
executing==2.2.0
feedparser==6.0.11
filelock==3.18.0
fonttools==4.55.3
fs==2.4.16
galternatives==1.0.10
html2text==2024.2.26
html5-parser==0.4.12
html5lib-modern==1.2
html5rdf==1.2.1
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
inflate64==1.0.1+ds
inflect==7.3.1
iotop==0.6
ipython==8.30.0
isodate==0.0.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.5
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
legacy-cgi==2.6.2
libvirt-python==11.1.0
louis==3.32.0
lxml==5.3.2
lxml_html_clean==0.4.1
lz4==4.4.0+dfsg
Mako==1.3.8.dev0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.10
menulibre==2.4.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.3
multivolumefile==0.2.3
natsort==8.0.2
nemo-compare==6.4.0
nemo-terminal==4.8.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.2.1
notify2==0.3.1
numpy==2.2.6
oauthlib==3.2.2
olefile==0.47
orjson==3.10.7
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.3.0
pipx==1.7.1
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.22.0+dfsg
pyasyncore==1.0.2
pybcj==1.0.3+ds
pycairo==1.27.0
pychm==0.8.6
pycparser==2.23
pycryptodomex==3.20.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
pylibacl==0.7.2
pyliblo3==0.16.2
pyparsing==3.1.2
pyppmd==1.1.1+ds
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6_sip==13.10.0
pyserial==3.5
python-apt==3.0.0
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
python-markdown-math==0.8
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.2
pyzstd==0.16.2
rdflib==7.1.1
rdiff-backup==2.2.6
referencing==0.35.1
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
Routes==2.5.1
rpds-py==0.21.0
scipy==1.14.1
scour==0.38.2
setuptools==75.8.0
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
soupsieve==2.6
ssh-import-id==5.11
stack-data==0.6.3
svgpathtools==1.7.1
svgwrite==1.4.3
sympy==1.13.3
systemd-python==235
terminator==2.1.4
texttable==1.7.0
tinycss2==1.4.0
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.2
types-aiofiles==24.1
types-antlr4-python3-runtime==4.13
types-assertpy==1.1
types-atheris==2.3
types-aws-xray-sdk==2.14
types-beautifulsoup4==4.12
types-bleach==6.2
types-boltons==24.1
types-braintree==4.31
types-cachetools==5.5
types-caldav==1.3
types-capturer==3.0
types-cffi==1.16
types-chevron==0.14
types-click-default-group==1.2
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.8
types-corus==0.10
types-croniter==5.0.1
types-dateparser==1.2
types-decorator==5.1
types-defusedxml==0.7
types-Deprecated==1.2.15
types-docker==7.1
types-dockerfile-parse==2.0
types-docutils==0.21
types-editdistance==0.8
types-entrypoints==0.4
types-ExifRead==3.0
types-fanstatic==1.4
types-first==2.0
types-flake8==7.1
types-flake8-bugbear==24.12.12
types-flake8-builtins==2.5
types-flake8-docstrings==1.7
types-flake8-rst-docstrings==0.3
types-flake8-simplify==0.21
types-flake8-typing-imports==1.16
types-Flask-Cors==5.0
types-Flask-Migrate==4.0
types-Flask-SocketIO==5.4
types-fpdf2==2.8.2
types-gdb==15.0
types-gevent==24.11
types-google-cloud-ndb==2.3
types-greenlet==3.1
types-hdbcli==2.23
types-html5lib==1.1
types-httplib2==0.22
types-humanfriendly==10.0
types-hvac==2.3
types-ibm-db==3.2.4
types-icalendar==6.1
types-influxdb-client==1.45
types-inifile==0.4
types-JACK-Client==0.5
types-Jetson.GPIO==2.1
types-jmespath==1.0
types-jsonschema==4.23
types-jwcrypto==1.5
types-keyboard==0.13
types-ldap3==2.9
types-libsass==0.23
types-lupa==2.2
types-lzstring==1.0
types-m3u8==6.0
types-Markdown==3.7
types-mock==5.1
types-mypy-extensions==1.0
types-mysqlclient==2.2
types-nanoid==2.0.0
types-netaddr==1.3
types-netifaces==0.11
types-networkx==3.4.2
types-oauthlib==3.2
types-objgraph==3.6
types-olefile==0.47
types-openpyxl==3.1.5
types-opentracing==2.4
types-paramiko==3.5
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.17.8
types-pep8-naming==0.14
types-pexpect==4.9
types-pika-ts==1.3
types-polib==1.2
types-portpicker==1.6
types-protobuf==5.29.1
types-psutil==6.1
types-psycopg2==2.9.10
types-pyasn1==0.6
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycocotools==2.0
types-pycurl==7.45.4
types-pyfarmhash==0.4
types-pyflakes==3.2
types-pygit2==1.15
types-Pygments==2.18
types-pyinstaller==6.11
types-pyjks==20.0
types-PyMySQL==1.1
types-pynput==1.7.7
types-pyOpenSSL==24.1
types-pyRFC3339==2.0.1
types-PyScreeze==1.0.1
types-pyserial==3.5
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==3.2
types-python-datemath==3.0.1
types-python-dateutil==2.9
types-python-http-client==3.3.7
types-python-jenkins==1.8
types-python-jose==3.3
types-python-nmap==0.7
types-python-xlib==0.33
types-pytz==2024.2
types-pywin32==308
types-pyxdg==0.28
types-PyYAML==6.0
types-qrbill==1.1
types-qrcode==8.0
types-regex==2024.11.6
types-reportlab==4.2.5
types-requests==2.32
types-requests-oauthlib==2.0
types-retry==0.9
types-RPi.GPIO==0.7
types-s2clientprotocol==5
types-seaborn==0.13.2
types-Send2Trash==1.8
types-setuptools==75.6
types-shapely==2.0
types-simplejson==3.19
types-singledispatch==4.1
types-six==1.17
types-slumber==0.7
types-str2bool==1.1
types-tabulate==0.9
types-tensorflow==2.18.0
types-TgCrypto==1.2
types-toml==0.10
types-toposort==1.10
types-tqdm==4.67
types-translationstring==1.4
types-tree-sitter-languages==1.10
types-ttkthemes==3.2
types-ujson==5.10
types-unidiff==0.7
types-untangle==1.2
types-usersettings==1.1
types-uWSGI==2.0
types-vobject==0.9.9
types-waitress==3.0.1
types-WebOb==1.8
types-whatthepatch==1.0
types-workalendar==17.0
types-WTForms==3.2.1
types-wurlitzer==3.1
types-xdgenvpy==3.0
types-xmltodict==0.14
types-zstd==1.5
types-zxcvbn==4.4
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.17.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.3.0
usb-creator==0.3.16
userpath==1.9.2
vboxapi==1.0
virtualenv==20.29.1
wadllib==2.0.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.9
wheel==0.45.1
wxPython==4.2.2
xdg==5
xkit==0.0.0
xlrd==2.0.1
xxhash==0.0.0
zeroconf==0.146.1
zipp==3.21.0
